+++ plone/app/portlets/manager.py
@@ -7,15 +7,13 @@ from zope.publisher.interfaces.browser import IBrowserView
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 from Acquisition import Explicit, aq_inner, aq_acquire
-from Products.CMFCore.utils import getToolByName
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from ZODB.POSException import ConflictError
 
-from plone.portlets.interfaces import IPortletRenderer, ILocalPortletAssignable
+from plone.portlets.interfaces import IPortletRenderer
 from plone.portlets.manager import PortletManagerRenderer as BasePortletManagerRenderer
 from plone.app.portlets.interfaces import IColumn
 from plone.app.portlets.interfaces import IDashboard
-#from plone.app.layout.navigation.defaultpage import isDefaultPage
 
 logger = logging.getLogger('portlets')
 
@@ -29,7 +27,7 @@ class PortletManagerRenderer(BasePortletManagerRenderer, Explicit):
         data object.
         """
         portlet = getMultiAdapter((self.context, self.request, self.__parent__,
-                                        self.manager, data, ), IPortletRenderer)
+                                  self.manager, data, ), IPortletRenderer)
         return portlet.__of__(self.context)
 
 
@@ -47,7 +45,8 @@ class ColumnPortletManagerRenderer(PortletManagerRenderer):
         """If context is a default-page, return URL of folder, else
         return URL of context.
         """
-        return str(getMultiAdapter((self._context(), self.request, ), name=u'absolute_url'))
+        return str(getMultiAdapter((self._context(), self.request, ),
+                                   name=u'absolute_url'))
 
     def safe_render(self, portlet_renderer):
         try: