+++ CHANGES.rst
@@ -10,6 +10,11 @@ Changelog
 3.0.5 (2015-05-12)
 ------------------
 
+- Convert manage-portlets.js into a pattern and make improvements on
+  using the manage portlets infrastructure
+  [vangheem]
+
+
 - Supress a ZopeTestCase warning.
   This fixes https://github.com/plone/Products.CMFPlone/issues/501
   [gforcada]